requirements.txt
pillow
tqdm
numpy
scikit-image
opencv-python